fix(inventory): replace bool with int values in ipfixprobe config
Output link and direction are expected to be either 0 or 1.
Boolean literals cause crash of ipfixprobe instances during start,
so there are no flows as a result.

Maybe we should modify `ansible/roles/pandda_conf/files/instance_ipfixprobe.conf.j2`
to automatically convert booleans to integers to prevent this kind
of errors altogether.